+++ hier_config/constructors.py
@@ -26,6 +26,7 @@ from .platforms.hp_procurve.view import HConfigViewHPProcurve
 from .platforms.huawei_vrp.driver import HConfigDriverHuaweiVrp
 from .platforms.juniper_junos.driver import HConfigDriverJuniperJUNOS
 from .platforms.nokia_srl.driver import HConfigDriverNokiaSRL
+from .platforms.ruckus_fastiron.driver import HConfigDriverRuckusFastIron
 from .platforms.view_base import HConfigViewBase
 from .platforms.vyos.driver import HConfigDriverVYOS
 from .root import HConfig
@@ -48,6 +49,7 @@ def get_hconfig_driver(platform: Platform) -> HConfigDriverBase:
         Platform.HUAWEI_VRP: HConfigDriverHuaweiVrp,
         Platform.JUNIPER_JUNOS: HConfigDriverJuniperJUNOS,
         Platform.NOKIA_SRL: HConfigDriverNokiaSRL,
+        Platform.RUCKUS_FASTIRON: HConfigDriverRuckusFastIron,
         Platform.VYOS: HConfigDriverVYOS,
     }
     driver_cls = platform_drivers.get(platform)